+++ sky_pointer.h
@@ -73,8 +73,8 @@ class SkyPointer {
 	uint8_t isLaserOn(void);    // Check the laser state
         void setLaserTimeout(uint32_t);
         void home();                // Find home position
-        void move(int16_t az, int16_t alt, uint8_t speed);
-        void goTo(uint16_t az, uint16_t alt, uint8_t speed);
+        void move(int16_t az, int16_t alt, uint16_t speed);
+        void goTo(uint16_t az, uint16_t alt, uint16_t speed);
         void getPos(uint16_t *az, uint16_t *alt);
         void stop();                // Stop both motors
         void releaseMotors();       // Disable the motor drivers
